Jsb bridge ready to review
