+++ cocos/platform/ohos/libcocos/src/main/java/com/cocos/lib/JsbBridge.java
@@ -22,14 +22,37 @@
  */
 package com.cocos.lib;
 
-import org.json.JSONObject;
+public class JsbBridge {
+    public interface ICallback{
+        /**
+         * Applies this callback to the given argument.
+         *
+         * @param arg0 as input
+         * @param arg1 as input
+         */
+        void onScript(String arg0, String arg1);
+    }
+    private static ICallback callback;
 
-public interface IMethod{
+    private static boolean callByScript(String arg0, String arg1){
+        if(JsbBridge.callback == null)
+            return false;
+        callback.onScript(arg0, arg1);
+        return true;
+    }
+
+    /**Add a callback which you would like to apply
+     * @param f ICallback, the method which will be actually applied. multiple calls will override
+     * */
+    public static void setCallback(ICallback f){
+        JsbBridge.callback = f;
+    }
     /**
-     * Applies this function to the given argument.
-     *
-     * @param t the function argument
+     * Java dispatch Js event, use native c++ code
+     * @param arg0 input values
      */
-    public void apply(String t);
-
+    public static native void sendToScript(String arg0, String arg1);
+    public static void sendToScript(String arg0){
+        sendToScript(arg0, null);
+    }
 }
